1) read the alignments in in-place 2) separatae Sample and bootstrap sample

+++ include/Reference.hpp
@@ -5,17 +5,17 @@
 #include <array>
 
 struct Grouping {
-  std::vector<unsigned short> indicators;
-  std::vector<short unsigned> sizes;
+  std::vector<uint32_t> indicators;
+  std::vector<uint16_t> sizes;
   std::vector<std::array<double, 2>> bb_params;
-  unsigned short n_groups;
+  uint32_t n_groups;
 };
 
 class Reference {
 public:
   Grouping grouping;
   std::vector<std::string> group_names;
-  unsigned n_refs;
+  uint32_t n_refs;
   
   void calculate_bb_parameters(double params[2]);
 };
